Add separate metric for counting published vocabularies and terms
